Replace ProgressBars2 in survey back office and ungrouped report builder widgets
