+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/MultipleChoice/index.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
 
-import { Box, colors, Image } from '@citizenlab/cl2-component-library';
+import { colors } from '@citizenlab/cl2-component-library';
 
 import { Answer, AnswerMultilocs } from 'api/survey_results/types';
 
-import useLocalize from 'hooks/useLocalize';
-
-import ProgressBars2 from 'components/admin/Graphs/ProgressBars2';
+import SurveyBars from 'components/admin/Graphs/SurveyBars';
 
 interface Props {
   multipleChoiceAnswers: Answer[];
   totalResponses: number;
-  multilocs?: AnswerMultilocs;
+  multilocs: AnswerMultilocs;
 }
 
 const COLOR_SCHEME = [colors.primary];
@@ -21,44 +19,14 @@ const MultipleChoice = ({
   totalResponses,
   multilocs,
 }: Props) => {
-  const localize = useLocalize();
-
   return (
-    <>
-      {multipleChoiceAnswers.map(({ answer, count }, index) => {
-        if (!multilocs || answer === null) return null;
-
-        const label = localize(multilocs.answer[answer].title_multiloc);
-        const image = multilocs.answer[answer].image;
-
-        return (
-          <Box
-            key={index}
-            maxWidth="524px"
-            display="flex"
-            alignItems="flex-end"
-            justifyContent="center"
-          >
-            {image?.small && (
-              <Box mr="12px">
-                <Image
-                  width="48px"
-                  height="48px"
-                  src={image.small}
-                  alt={label}
-                />
-              </Box>
-            )}
-            <ProgressBars2
-              values={[count]}
-              total={totalResponses}
-              colorScheme={COLOR_SCHEME}
-              label={label}
-            />
-          </Box>
-        );
-      })}
-    </>
+    <SurveyBars
+      grouped={false}
+      answers={multipleChoiceAnswers}
+      totalResponses={totalResponses}
+      multilocs={multilocs}
+      colorScheme={COLOR_SCHEME}
+    />
   );
 };
