+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/SurveyQuestionResultWidget/Question/UngroupedBars.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
 
-import { Box, Image } from '@citizenlab/cl2-component-library';
+import { Box } from '@citizenlab/cl2-component-library';
 
 import { AttributesUngrouped } from 'api/graph_data_units/responseTypes';
 
-import useLocalize from 'hooks/useLocalize';
-
-import ProgressBars2 from 'components/admin/Graphs/ProgressBars2';
 import { DEFAULT_CATEGORICAL_COLORS } from 'components/admin/Graphs/styling';
-
-import { useIntl } from 'utils/cl-intl';
-
-import messages from '../messages';
+import SurveyBars from 'components/admin/Graphs/SurveyBars';
 
 interface Props {
   attributes: AttributesUngrouped;
@@ -20,47 +14,15 @@ interface Props {
 const COLOR_SCHEME = [DEFAULT_CATEGORICAL_COLORS[0]];
 
 const UngroupedBars = ({ attributes }: Props) => {
-  const { formatMessage } = useIntl();
-  const localize = useLocalize();
-
   return (
     <Box className="e2e-survey-question-ungrouped-bars">
-      {attributes.answers.map(({ answer, count }, index) => {
-        const image = answer
-          ? attributes.multilocs.answer[answer].image
-          : undefined;
-        const label =
-          answer === null
-            ? formatMessage(messages.noAnswer)
-            : localize(attributes.multilocs.answer[answer].title_multiloc);
-
-        return (
-          <Box
-            key={index}
-            maxWidth="524px"
-            display="flex"
-            alignItems="flex-end"
-            justifyContent="center"
-          >
-            {image?.small && (
-              <Box mr="12px">
-                <Image
-                  width="48px"
-                  height="48px"
-                  src={image.small}
-                  alt={label}
-                />
-              </Box>
-            )}
-            <ProgressBars2
-              values={[count]}
-              total={attributes.totalPickCount}
-              colorScheme={COLOR_SCHEME}
-              label={label}
-            />
-          </Box>
-        );
-      })}
+      <SurveyBars
+        grouped={false}
+        answers={attributes.answers}
+        totalResponses={attributes.totalPickCount}
+        multilocs={attributes.multilocs}
+        colorScheme={COLOR_SCHEME}
+      />
     </Box>
   );
 };
