+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/_deprecated/SurveyResultsWidget/FormResultsQuestion.tsx
@@ -40,7 +40,7 @@ const FormResultsQuestion = ({
           totalSubmissions={totalSubmissions}
           totalResponses={questionResponseCount}
         />
-        {isMultipleChoiceAndHasAnswers && (
+        {isMultipleChoiceAndHasAnswers && multilocs && (
           <MultipleChoice
             multipleChoiceAnswers={answers}
             totalResponses={questionResponseCount}
